fix lang on html tag

+++ config/RemovePages.js
@@ -5,7 +5,12 @@ fs.readdir(path.resolve(__dirname, '../pages'), (err, files) => {
 	if (err) throw err
 	if (files.includes('index.js')) {
 		files.forEach(file => {
-			if (file !== '404.js' && file !== '_app.js' && file !== '_error.js') {
+			if (
+				file !== '404.js' &&
+				file !== '_app.js' &&
+				file !== '_error.js' &&
+				file !== '_document.js'
+			) {
 				fs.rm(
 					path.resolve(__dirname, '../pages', file),
 					{ recursive: true, force: true },
